requirements.txt
discord.py>=2.3.2
psycopg2-binary
python-dotenv
asyncpg
pytest